requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudinary==1.29.0
config==0.3.9
dj-database-url==0.5.0
Django==4.0.5
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.2.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.28.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.9
whitenoise==6.2.0
